@@ -1,9 +1,8 @@
 {
   "name": "rmwc-main",
   "author": "James Friedman",
-  "version": "5.1.8",
+  "version": "5.2.0-alpha.1",
   "private": true,
-  "sideEffects": false,
   "homepage": "https://jamesmfriedman.github.io/rmwc/",
   "description": "RMWC: A React wrapper for Material Design (Web) Components",
   "license": "MIT",
@@ -22,9 +21,9 @@
     "prop-types": "^15.6.2"
   },
   "scripts": {
-    "install": "lerna bootstrap",
-    "start": "react-app-rewired start",
-    "build:docsite": "react-app-rewired build && rm -R -f docs && mv build docs",
+    "install": "lerna bootstrap --force-local --hoist",
+    "start": "export NODE_OPTIONS='--max-old-space-size=4096' && react-app-rewired start",
+    "build:docsite": "export NODE_OPTIONS='--max-old-space-size=4096' && react-app-rewired build && rm -R -f docs && mv build docs",
     "build:docs": "npm run clean && npm run docgen && npm run build:docsite",
     "build:code": "npm run clean && node scripts/build-code.js",
     "build:lib": "node scripts/build-lib.js",
@@ -33,8 +32,13 @@
     "test": "node scripts/test.js",
     "test:watch": "react-app-rewired test --env=jsdom",
     "test:coverage": "react-app-rewired test --env=jsdom --coverage",
-    "release": "npm run clean && lerna version && lerna run publish:dist --parallel --no-bail && npm run clean",
-    "release:force": "npm run clean && lerna version --force-publish && lerna run publish:dist --parallel --no-bail && npm run clean",
+    "release:patch": "lerna publish patch --npm-tag latest",
+    "release:minor": "lerna publish minor --npm-tag latest",
+    "release:major": "lerna publish major --npm-tag latest",
+    "alpha": "lerna publish --npm-tag next",
+    "alpha:patch": "lerna publish prepatch --npm-tag next",
+    "alpha:minor": "lerna publish preminor --npm-tag next",
+    "alpha:major": "lerna publish premajor --npm-tag next",
     "clean": "node scripts/clean.js",
     "docgen": "node scripts/docgen.js",
     "update-npmignore": "lerna exec -- cp \\$LERNA_ROOT_PATH/.npmignore .npmignore",
@@ -47,19 +51,22 @@
     "flow": "flow"
   },
   "devDependencies": {
-    "@storybook/addon-actions": "^4.1.11",
-    "@storybook/addon-knobs": "^4.1.11",
-    "@storybook/react": "^4.1.11",
+    "@storybook/addon-actions": "^5.0.0",
+    "@storybook/addon-knobs": "^5.0.0",
+    "@storybook/react": "^5.0.0",
     "@types/classnames": "^2.2.7",
     "@types/enzyme": "^3.1.15",
     "@types/enzyme-adapter-react-16": "^1.0.3",
-    "@types/jest": "^23.3.12",
+    "@types/jest": "^24.0.9",
     "@types/node": "^10.12.18",
     "@types/react": "^16.8.3",
     "@types/react-dom": "^16.0.11",
     "@types/react-router": "^4.4.4",
     "@types/react-router-dom": "^4.3.1",
-    "babel-loader": "^8.0.5",
+    "@types/storybook__addon-actions": "^3.4.2",
+    "@types/storybook__addon-knobs": "^4.0.2",
+    "@types/storybook__react": "^4.0.1",
+    "babel-loader": "8.0.5",
     "copyfiles": "^2.1.0",
     "enzyme": "^3.8.0",
     "enzyme-adapter-react-16": "1.8.0",
@@ -75,12 +82,12 @@
     "react-hot-loader": "^4.6.3",
     "react-markdown-loader": "git+https://github.com/jamesmfriedman/react-markdown-loader.git#master",
     "react-router-dom": "^4.3.1",
-    "react-scripts": "2.1.5",
+    "react-scripts": "^2.1.8",
     "react-styleguidist": "^8.0.6",
     "tslint": "^5.12.1",
     "typedoc": "^0.14.2",
-    "typescript": "^3.2.2",
-    "webpack": "^4.28.3"
+    "typescript": "^3.3.3333",
+    "webpack": "4.28.3"
   },
   "eslintIgnore": [
     "node_modules"
