@@ -1,6 +1,6 @@
 import * as RMWC from '@rmwc/types';
 import * as React from 'react';
-//@ts-ignore
+
 import { MDCChipFoundation } from '@material/chips';
 
 import {
@@ -10,7 +10,7 @@ import {
   deprecationWarning
 } from '@rmwc/base';
 import { withRipple } from '@rmwc/ripple';
-import { Icon, IconProps, IconPropT } from '@rmwc/icon';
+import { Icon, IconProps } from '@rmwc/icon';
 
 export interface ChipProps {
   /** Text for your Chip. */
@@ -18,9 +18,9 @@ export interface ChipProps {
   /** makes the Chip appear selected. */
   selected?: boolean;
   /** Instance of an Icon Component. */
-  icon?: IconPropT;
+  icon?: RMWC.IconPropT;
   /** Instance of an Icon Component. */
-  trailingIcon?: IconPropT;
+  trailingIcon?: RMWC.IconPropT;
   /** An optional chip ID that will be included in callback evt.detail. If this is not passed, RMWC will attempt to use the "key" prop if present.  */
   id?: string;
   /** Includes an optional checkmark for the chips selected state. */
@@ -56,7 +56,10 @@ const ChipRoot = withRipple({})(
 );
 
 /** A Chip component. */
-export class Chip extends FoundationComponent<ChipProps & DeprecatedChipProps> {
+export class Chip extends FoundationComponent<
+  MDCChipFoundation,
+  ChipProps & DeprecatedChipProps
+> {
   static displayName = 'Chip';
 
   private root = this.createElement('root');
@@ -142,13 +145,15 @@ export class Chip extends FoundationComponent<ChipProps & DeprecatedChipProps> {
     );
   }
 
-  handleInteraction(evt: React.MouseEvent & React.KeyboardEvent) {
+  handleInteraction(
+    evt: React.MouseEvent & React.KeyboardEvent & MouseEvent & KeyboardEvent
+  ) {
     evt.type === 'click' && this.props.onClick && this.props.onClick(evt);
     evt.type === 'keydown' && this.props.onKeyDown && this.props.onKeyDown(evt);
     return this.foundation.handleInteraction(evt);
   }
 
-  handleTransitionEnd(evt: React.TransitionEvent) {
+  handleTransitionEnd(evt: React.TransitionEvent & TransitionEvent) {
     this.foundation.handleTransitionEnd(evt);
   }
 
