@@ -3,63 +3,20 @@ import * as React from 'react';
 import { withProviderContext, WithProviderContext } from '@rmwc/provider';
 import { componentFactory, classNames, deprecationWarning } from '@rmwc/base';
 
-type IconElementT = React.ReactNode;
-export type IconSizeT = 'xsmall' | 'small' | 'medium' | 'large' | 'xlarge';
-
-export type IconStrategyT =
-  | 'auto'
-  | 'ligature'
-  | 'className'
-  | 'url'
-  | 'component'
-  | 'custom';
-
-export interface IconOptions {
-  icon: IconElementT;
-  /**
-   * Handle multiple methods of embedding an icon.
-   * 'ligature' uses ligature style embedding like material-icons,
-   * 'className' adds a class onto the element for libraries like glyphicons and ion icons,
-   * 'url' will load a remote image, and
-   * 'component' will render content as children like SVGs or any other React node.
-   * 'custom' allows you to specify your own render prop.
-   * If not set, 'auto' will be used or the defaults set inside of RMWCProvider.
-   * */
-  strategy?: IconStrategyT;
-  /**
-   * A className prefix to use when using css font icons that use prefixes,
-   * i.e. font-awesome-, ion-, glyphicons-.
-   * This only applies when using the 'className' strategy.
-   **/
-  prefix?: string;
-  /** A base className for the icon namespace, i.e. material-icons. */
-  basename?: string;
-  /** A render function to use when using the 'custom' strategy. */
-  render?: (
-    props: { content: IconElementT; className: string }
-  ) => React.ReactNode;
-  /** A size to render the icon  */
-  size?: IconSizeT;
-  /** Additional props */
-  [key: string]: any;
-}
-
-export type IconPropT = IconElementT | IconOptions;
-
 export interface IconProps extends DeprecatedIconProps {
   /** The icon to use. This can be a string for a font icon, a url, or whatever the selected strategy needs. */
-  icon?: IconPropT;
+  icon?: RMWC.IconPropT;
 }
 
 export interface DeprecatedIconProps {
   /** DEPRECATED: Additional Icon Options. See the Icon component documentation. */
-  iconOptions?: IconOptions;
+  iconOptions?: RMWC.IconOptions;
 }
 
 /**
  * Given content, tries to figure out an appropriate strategy for it
  */
-const processAutoStrategy = (content: React.ReactNode): IconStrategyT => {
+const processAutoStrategy = (content: React.ReactNode): RMWC.IconStrategyT => {
   // check for URLS
   if (typeof content === 'string' && content.includes('/')) {
     return 'url';
@@ -130,13 +87,13 @@ const iconRenderMap: {
   auto: undefined
 };
 
-const buildIconOptions = (icon?: IconPropT) => {
+const buildIconOptions = (icon?: RMWC.IconPropT) => {
   if (React.isValidElement(icon) || (icon && typeof icon !== 'object')) {
     return {
       icon
     };
   }
-  return icon as IconOptions;
+  return icon as RMWC.IconOptions;
 };
 
 const IconRoot = componentFactory({ displayName: 'IconRoot', tag: 'i' });
@@ -177,7 +134,7 @@ export const Icon = withProviderContext()(
       render,
       size,
       ...optionsRest
-    }: IconOptions = {
+    }: RMWC.IconOptions = {
       ...buildIconOptions(icon),
       ...deprecatedIconOption
     };
