@@ -1,7 +1,7 @@
 import * as RMWC from '@rmwc/types';
 import * as React from 'react';
 
-// @ts-ignore
+import { EventType, SpecificEventListener } from '@material/base/types';
 import { MDCSliderFoundation } from '@material/slider';
 import { componentFactory, FoundationComponent, debounce } from '@rmwc/base';
 
@@ -119,7 +119,11 @@ type SliderState = {
   trackMarkersCount: number;
 };
 
-export class Slider extends FoundationComponent<SliderProps, SliderState> {
+export class Slider extends FoundationComponent<
+  MDCSliderFoundation,
+  SliderProps,
+  SliderState
+> {
   state = {
     trackMarkersCount: 0,
     pinContainerStyle: {}
@@ -134,11 +138,11 @@ export class Slider extends FoundationComponent<SliderProps, SliderState> {
     super.componentDidMount();
     // Fixes an issue where synthetic events were being
     // accessed in the Foundation and causing an error
-    const existinghandleDown_ = this.foundation.handleDown_.bind(
+    const existinghandleDown_ = (this.foundation as any).handleDown_.bind(
       this.foundation
     );
 
-    this.foundation.handleDown_ = (evt: React.SyntheticEvent<any>) => {
+    (this.foundation as any).handleDown_ = (evt: React.SyntheticEvent<any>) => {
       evt.persist();
       existinghandleDown_(evt);
     };
@@ -195,22 +199,22 @@ export class Slider extends FoundationComponent<SliderProps, SliderState> {
   }
 
   get discrete(): boolean {
-    return !!(this.foundation && this.foundation.isDiscrete_);
+    return !!(this.foundation && (this.foundation as any).isDiscrete_);
   }
 
   set discrete(isDiscrete: boolean) {
     if (this.foundation) {
-      this.foundation.isDiscrete_ = isDiscrete;
+      (this.foundation as any).isDiscrete_ = isDiscrete;
     }
   }
 
   get displayMarkers(): boolean {
-    return !!this.foundation && this.foundation.hasTrackMarker_;
+    return !!this.foundation && (this.foundation as any).hasTrackMarker_;
   }
 
   set displayMarkers(isDisplayMarkers: boolean) {
     if (this.foundation) {
-      this.foundation.hasTrackMarker_ = isDisplayMarkers;
+      (this.foundation as any).hasTrackMarker_ = isDisplayMarkers;
     }
   }
 
@@ -271,43 +275,62 @@ export class Slider extends FoundationComponent<SliderProps, SliderState> {
       hasClass: (className: string) => this.root.hasClass(className),
       addClass: (className: string) => this.root.addClass(className),
       removeClass: (className: string) => this.root.removeClass(className),
-      getAttribute: (name: string) => this.root.getProp(name as any),
+      getAttribute: (name: string) =>
+        this.root.getProp(name as any) as string | null,
       setAttribute: debounce(
         (name: string, value: any) => this.root.setProp(name as any, value),
         300
       ),
       removeAttribute: (name: string) => this.root.removeProp(name as any),
       computeBoundingRect: () =>
-        this.root.ref && this.root.ref.getBoundingClientRect(),
-      getTabIndex: () => this.root.ref && this.root.ref.tabIndex,
-      registerInteractionHandler: (type: string, handler: () => void) => {
-        this.root.addEventListener(type, handler);
+        this.root.ref
+          ? this.root.ref.getBoundingClientRect()
+          : ({} as ClientRect),
+      getTabIndex: () => (this.root.ref ? this.root.ref.tabIndex : 0),
+      registerInteractionHandler: <K extends EventType>(
+        evtType: K,
+        handler: SpecificEventListener<K>
+      ): void => {
+        this.root.addEventListener(evtType, handler);
       },
-      deregisterInteractionHandler: (type: string, handler: () => void) => {
-        this.root.removeEventListener(type, handler);
+      deregisterInteractionHandler: <K extends EventType>(
+        evtType: K,
+        handler: SpecificEventListener<K>
+      ): void => {
+        this.root.removeEventListener(evtType, handler);
       },
-      registerThumbContainerInteractionHandler: (
-        type: string,
-        handler: () => void
-      ) => {
-        this.thumbContainer.addEventListener(type, handler);
+      registerThumbContainerInteractionHandler: <K extends EventType>(
+        evtType: K,
+        handler: SpecificEventListener<K>
+      ): void => {
+        this.thumbContainer.addEventListener(evtType, handler);
       },
-      deregisterThumbContainerInteractionHandler: (
-        type: string,
-        handler: () => void
-      ) => {
-        this.thumbContainer.removeEventListener(type, handler);
+      deregisterThumbContainerInteractionHandler: <K extends EventType>(
+        evtType: K,
+        handler: SpecificEventListener<K>
+      ): void => {
+        this.thumbContainer.removeEventListener(evtType, handler);
       },
-      registerBodyInteractionHandler: (type: string, handler: () => void) => {
-        document.body && document.body.addEventListener(type, handler);
+      registerBodyInteractionHandler: <K extends EventType>(
+        evtType: K,
+        handler: SpecificEventListener<K>
+      ): void => {
+        document.body && document.body.addEventListener(evtType, handler);
       },
-      deregisterBodyInteractionHandler: (type: string, handler: () => void) => {
-        document.body && document.body.removeEventListener(type, handler);
+      deregisterBodyInteractionHandler: <K extends EventType>(
+        evtType: K,
+        handler: SpecificEventListener<K>
+      ): void => {
+        document.body && document.body.removeEventListener(evtType, handler);
       },
-      registerResizeHandler: (handler: () => void) => {
+      registerResizeHandler: (
+        handler: SpecificEventListener<'resize'>
+      ): void => {
         window.addEventListener('resize', handler);
       },
-      deregisterResizeHandler: (handler: () => void) => {
+      deregisterResizeHandler: (
+        handler: SpecificEventListener<'resize'>
+      ): void => {
         window.removeEventListener('resize', handler);
       },
       notifyInput: () => {
@@ -322,7 +345,7 @@ export class Slider extends FoundationComponent<SliderProps, SliderState> {
       setTrackStyleProperty: (propertyName: string, value: any) => {
         this.track && this.track.style.setProperty(propertyName, value);
       },
-      setMarkerValue: (value: string) => {
+      setMarkerValue: (value: number) => {
         this.sliderPin.setProp('value', value);
       },
       appendTrackMarkers: (numMarkers: number) => {
@@ -334,15 +357,15 @@ export class Slider extends FoundationComponent<SliderProps, SliderState> {
       setLastTrackMarkersStyleProperty: (propertyName: string, value: any) => {
         if (this.root.ref) {
           // We remove and append new nodes, thus, the last track marker must be dynamically found.
-          const lastTrackMarker = this.root.ref.querySelector(
-            MDCSliderFoundation.strings.LAST_TRACKMARKER_SELECTOR
+          const lastTrackMarker = this.root.ref.querySelector<HTMLElement>(
+            MDCSliderFoundation.strings.LAST_TRACK_MARKER_SELECTOR
           );
           lastTrackMarker &&
             lastTrackMarker.style.setProperty(propertyName, value);
         }
       },
       isRTL: () =>
-        this.root.ref && getComputedStyle(this.root.ref).direction === 'rtl'
+        !!this.root.ref && getComputedStyle(this.root.ref).direction === 'rtl'
     });
   }
 
