@@ -1,7 +1,5 @@
 import * as RMWC from '@rmwc/types';
 import * as React from 'react';
-
-// @ts-ignore
 import { MDCSwitchFoundation } from '@material/switch';
 import { componentFactory, classNames } from '@rmwc/base';
 import { withRipple } from '@rmwc/ripple';
@@ -61,7 +59,10 @@ const SwitchNativeControl = componentFactory<{}>({
   classNames: ['mdc-switch__native-control']
 });
 
-export class Switch extends ToggleableFoundationComponent<SwitchProps> {
+export class Switch extends ToggleableFoundationComponent<
+  MDCSwitchFoundation,
+  SwitchProps
+> {
   static displayName = 'Switch';
   private root = this.createElement('root');
   private nativeControl = this.createElement<HTMLInputElement>('nativeControl');
@@ -74,7 +75,9 @@ export class Switch extends ToggleableFoundationComponent<SwitchProps> {
   componentDidMount() {
     super.componentDidMount();
     this.nativeControl.ref &&
-      this.foundation.updateCheckedStyling_(this.nativeControl.ref.checked);
+      (this.foundation as any).updateCheckedStyling_(
+        this.nativeControl.ref.checked
+      );
     this.nativeControl.ref &&
       this.foundation.setDisabled(this.nativeControl.ref.disabled);
   }
@@ -101,7 +104,7 @@ export class Switch extends ToggleableFoundationComponent<SwitchProps> {
       !!prevProps &&
       props.checked !== prevProps.checked
     ) {
-      this.foundation.updateCheckedStyling_(props.checked);
+      (this.foundation as any).updateCheckedStyling_(props.checked);
     }
 
     if (
