@@ -4,16 +4,18 @@ import {
   MDCFadingTabIndicatorFoundation,
   MDCTabIndicatorFoundation,
   MDCSlidingTabIndicatorFoundation
-  // @ts-ignore
 } from '@material/tab-indicator';
 import { FoundationComponent } from '@rmwc/base';
 
-export class TabIndicator extends FoundationComponent<{}> {
+export class TabIndicator extends FoundationComponent<
+  MDCSlidingTabIndicatorFoundation,
+  {}
+> {
   private root = this.createElement('root');
   private content = this.createElement('content');
 
   getDefaultFoundation() {
-    const adapter = /** @type {!MDCTabIndicatorAdapter} */ (Object.assign({
+    return new MDCSlidingTabIndicatorFoundation({
       addClass: (className: string) => {
         this.root.addClass(className);
       },
@@ -21,54 +23,17 @@ export class TabIndicator extends FoundationComponent<{}> {
         this.root.removeClass(className);
       },
       computeContentClientRect: () =>
-        this.content.ref && this.content.ref.getBoundingClientRect(),
+        this.content.ref
+          ? this.content.ref.getBoundingClientRect()
+          : ({} as ClientRect),
       setContentStyleProperty: (prop: string, value: string) => {
         this.content.setStyle(prop, value);
       }
-    }));
-
-    // if (this.root.hasClass(MDCTabIndicatorFoundation.cssClasses.FADE)) {
-    //   return new MDCFadingTabIndicatorFoundation(adapter);
-    // }
-
-    // Default to the sliding indicator
-    return new MDCSlidingTabIndicatorFoundation(adapter);
+    });
   }
 
-  activate(previousIndicatorClientRect: ClientRect) {
-    // Early exit if no indicator is present to handle cases where an indicator
-    // may be activated without a prior indicator state
-    if (!previousIndicatorClientRect) {
-      this.foundation.adapter_.addClass(
-        MDCTabIndicatorFoundation.cssClasses.ACTIVE
-      );
-      return;
-    }
-
-    const currentClientRect = this.computeContentClientRect();
-    const widthDelta =
-      previousIndicatorClientRect.width / currentClientRect.width;
-    const xPosition = previousIndicatorClientRect.left - currentClientRect.left;
-    this.foundation.adapter_.addClass(
-      MDCTabIndicatorFoundation.cssClasses.NO_TRANSITION
-    );
-    this.foundation.adapter_.setContentStyleProperty(
-      'transform',
-      `translateX(${xPosition}px) scaleX(${widthDelta})`
-    );
-
-    requestAnimationFrame(() => {
-      // Fixes an error of this executing after unmounting
-      if (!this.foundation) return;
-
-      this.foundation.adapter_.removeClass(
-        MDCTabIndicatorFoundation.cssClasses.NO_TRANSITION
-      );
-      this.foundation.adapter_.addClass(
-        MDCTabIndicatorFoundation.cssClasses.ACTIVE
-      );
-      this.foundation.adapter_.setContentStyleProperty('transform', '');
-    });
+  activate(previousIndicatorClientRect?: ClientRect) {
+    this.foundation.activate(previousIndicatorClientRect);
   }
 
   deactivate() {
